requirements.txt
python-dotenv
lxml
azure-core==1.38.2
azure-storage-blob==12.28.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
